+++ js/chatbot/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @gradio/chatbot
 
+## 0.24.3
+
+### Fixes
+
+- [#10724](https://github.com/gradio-app/gradio/pull/10724) [`667f8a1`](https://github.com/gradio-app/gradio/commit/667f8a1f820d46a4d87ad0f65b9e3b37c4de449f) - Fix Chatbot avatar image position.  Thanks @abidlabs!
+
+### Dependency updates
+
+- @gradio/video@0.14.4
+- @gradio/client@1.13.1
+- @gradio/wasm@0.17.4
+- @gradio/gallery@0.15.9
+- @gradio/upload@0.15.4
+- @gradio/image@0.21.4
+
 ## 0.24.2
 
 ### Dependency updates